fix(routing): relocate guard off scripts/ root + kill the qwen3.7-max silent default
CI root-scripts gate + the guard's own findings:
- routing_guard.py → scripts/ai_agent_bridge/ (root gate compliance);
  delegate.py gets a self-sufficient sys.path shim (it runs as a plain script).
- OPENCODE_DEFAULT_MODEL was openrouter/qwen/qwen3.7-max — the EXPENSIVE qwen
  as silent default for every ask-opencode call without --model (the user's
  qwen spend root cause, May-2026 78-turn burn class). Now gemma-4-31b-it
  (cheap documented surface lane, guard-allowed) with a regression test that
  the default passes the guard.
- test fixtures updated off the now-forbidden qwen pin.

X-Agent: claude-inline/routing-guard
