+++ scripts/ai_agent_bridge/routing_guard.py
