+++ tests/test_ask_opencode.py
@@ -7,20 +7,23 @@ import pytest
 from scripts.ai_agent_bridge._opencode import OPENCODE_DEFAULT_MODEL, _invoke_opencode
 
 
-def test_opencode_default_model_is_qwen_max():
-    assert OPENCODE_DEFAULT_MODEL == "openrouter/qwen/qwen3.7-max"
+def test_opencode_default_model_is_cheap_and_guard_allowed():
+    from scripts.ai_agent_bridge.routing_guard import assert_model_routing_allowed
+
+    assert OPENCODE_DEFAULT_MODEL == "openrouter/google/gemma-4-31b-it"
+    assert_model_routing_allowed(OPENCODE_DEFAULT_MODEL, context="default-model test")
 
 
 def test_invoke_opencode_constructs_correct_argv():
     with patch("scripts.ai_agent_bridge._opencode.shutil.which", return_value="/fake/opencode"):
         with patch("scripts.ai_agent_bridge._opencode.subprocess.run") as run_mock:
             run_mock.return_value = MagicMock(returncode=0, stdout="response", stderr="")
-            _invoke_opencode("hello", "openrouter/qwen/qwen3.7-max")
+            _invoke_opencode("hello", "openrouter/google/gemma-4-31b-it")
             argv = run_mock.call_args[0][0]
             assert argv[0] == "/fake/opencode"
             assert argv[1] == "run"
             assert "--model" in argv
-            assert "openrouter/qwen/qwen3.7-max" in argv
+            assert "openrouter/google/gemma-4-31b-it" in argv
             assert "hello" in argv
 
 
@@ -30,7 +33,7 @@ def test_invoke_opencode_attaches_file(tmp_path):
     with patch("scripts.ai_agent_bridge._opencode.shutil.which", return_value="/fake/opencode"):
         with patch("scripts.ai_agent_bridge._opencode.subprocess.run") as run_mock:
             run_mock.return_value = MagicMock(returncode=0, stdout="ok", stderr="")
-            _invoke_opencode("review", "openrouter/qwen/qwen3.7-max", data=str(data_file))
+            _invoke_opencode("review", "openrouter/google/gemma-4-31b-it", data=str(data_file))
             argv = run_mock.call_args[0][0]
             assert "--file" in argv
             # file path is in argv right after --file
@@ -42,4 +45,4 @@ def test_invoke_opencode_attaches_file(tmp_path):
 def test_invoke_opencode_raises_when_binary_missing():
     with patch("scripts.ai_agent_bridge._opencode.shutil.which", return_value=None):
         with pytest.raises(SystemExit, match="opencode CLI not found"):
-            _invoke_opencode("hello", "openrouter/qwen/qwen3.7-max")
+            _invoke_opencode("hello", "openrouter/google/gemma-4-31b-it")
